users must use leave conversation to remove themselves from a convo

+++ src/main/java/com/zfgc/services/pm/PmService.java
@@ -440,6 +440,11 @@ public class PmService extends AbstractService {
 			throw new ZfgcUnauthorizedException();
 		}
 		
+		//also verify that the user is not trying to remove themselves
+		if(zfgcUser.getUsersId() == remove.getUsersId()){
+			throw new ZfgcUnauthorizedException();
+		}
+		
 		pmConversationDataProvider.deleteConversationFromBox(convo, remove);
 	}
